move back locale to main.development
